docs(table): prose
