--- elements/pf-table/demo/demo.css
@@ -1,4 +1,0 @@
-
-section {
-  padding: 15px 30px;
-}
