+++ elements/pf-table/demo/expandable-rows-compound.html
@@ -14,11 +14,11 @@
     <pf-td>
       <a href="#">siemur/test-space</a>
     </pf-td>
-    <pf-td compound-expand="branches">
+    <pf-td compound-expand="ice-cream">
       <pf-icon set="fas" icon="code-branch"></pf-icon>
       10
     </pf-td>
-    <p slot="branches">Branches...</p>
+    <p slot="ice-cream">Branches...</p>
     <pf-td compound-expand="pull-requests">
       <pf-icon set="fas" icon="code-pull-request"></pf-icon>
       4
