--- elements/pf-table/demo/pf-table.js
@@ -1,1 +1,0 @@
-import '@patternfly/elements/pf-table/pf-table.js';
