+++ elements/pf-table/demo/sortable.html
@@ -1,6 +1,3 @@
-<link rel="stylesheet" href="demo.css"/>
-<script type="module" src="pf-table.js"></script>
-
 <pf-table>
   <pf-thead>
     <pf-tr>
@@ -70,3 +67,7 @@
   </pf-tr>
 </pf-table>
 
+<script type="module">
+  import '@patternfly/elements/pf-icon/pf-icon.js';
+  import '@patternfly/elements/pf-table/pf-table.js';
+</script>
